+++ R/write_codemeta.R
@@ -13,6 +13,7 @@
 #'  to package root. Default guess is current dir.
 #' @param id identifier for the package, e.g. a DOI (or other resolvable URL)
 #' @param force_update Update guessed fields even if they are defined in an existing codemeta.json file
+#' @param verbose Whether to print messages indicating opinions e.g. when DESCRIPTION has no URL. See \code{\link{give_opinions}}.
 #' @param ...  additional arguments to \code{\link{write_json}}
 #' @details If pkg is a codemeta object, the function will attempt to
 #'  update any fields it can guess (i.e. from the DESRIPTION file),
@@ -21,26 +22,52 @@
 #' @return writes out the codemeta.json file
 #' @export
 #'
-#' @importFrom jsonlite write_json
-#' @importFrom devtools use_build_ignore
-# upcoming devtools release will switch this to:
-# @importFrom usethis use_build_ignore
 #' @examples
-#' write_codemeta("codemeta", tempfile())
+#' write_codemeta("codemetar", tempfile())
+#' @details
+#'
+#' When creating and writing a codemeta.json for the first time,
+#' the function adds "codemeta.json" to .Rbuildignore and, if the
+#' project uses git, adds a pre-commit hook ensuring that if DESCRIPTION changes,
+#'  the codemeta.json will be updated as well unless the DESCRIPTION change is committed
+#'  with 'git commit --no-verify'.
 write_codemeta <- function(pkg = ".",
                            path = "codemeta.json",
                            root = ".",
                            id = NULL,
                            force_update =
                              getOption("codemeta_force_update", TRUE),
+                           verbose = TRUE,
                            ...) {
 
   if(length(pkg) <= 1 && file.exists(file.path(pkg, "DESCRIPTION"))){
-    devtools::use_build_ignore("codemeta.json", pkg = pkg)
-    ## usethis::use_build_ignore("codemeta.json", base_path = pkg)
+    usethis::use_build_ignore("codemeta.json")
+    # add the git pre-commit hook
+    # https://github.com/r-lib/usethis/blob/master/inst/templates/readme-rmd-pre-commit.sh#L1
+    # this is GPL-3 code
+    if (uses_git()) {
+      if(!file.exists(file.path(pkg, "codemeta.json"))){
+        message("* Adding a pre-commit git hook ensuring that codemeta.json will be synchronized with DESCRIPTION") # nolint
+        usethis::use_git_hook(
+          "pre-commit",
+          render_template("description-codemetajson-pre-commit.sh")
+        )
+        message(
+          "* Include the following code somewhere in R/ in your package, this way devtools::release() will remind you to update codemeta.json.\n",
+          '  release_questions <- function() "Have you updated codemeta.json with codemetar::write_codemeta()?"')
+      }
+
+    }
   }
   cm <- create_codemeta(pkg = pkg, root = root)
-  jsonlite::write_json(cm, path, pretty=TRUE, auto_unbox = TRUE, ...)
+  write_json(cm, path, pretty=TRUE, auto_unbox = TRUE, ...)
 
 }
 
+# from https://github.com/jeroen/jsonlite/blob/1f9e609e7d0ed702ede9c82aa5482ba08d5e5ab2/R/read_json.R#L22
+#' @param x an object to be serialized to JSON
+#' @param ... additional arguments passed to \link{toJSON} or \link{fromJSON}
+write_json <- function(x, path, ...) {
+  json <- jsonlite::toJSON(x, ...)
+  writeLines(json, path, useBytes = TRUE)
+}