+++ tests/testthat/test-parse_depends.R
@@ -2,23 +2,43 @@ testthat::context("parse_depends.R")
 
 testthat::test_that("Test the various cases for dependencies", {
 
-  a <- parse_depends(NULL)
-  testthat::expect_length(a, 0)
-  a <- parse_depends(deps = "a4")  # BIOC provider
-  testthat::expect_gt(length(a[[1]]), 1)
-  a <- parse_depends(deps = "httr") # CRAN provider
-  testthat::expect_gt(length(a[[1]]), 1)
-  a <- parse_depends(deps = "R")
-  testthat::expect_equal(a[[1]]$name, "R")
-  a <- parse_depends(deps = "not-a-package")
+  testthat::expect_error(format_depend(NULL))
+  a <- format_depend(package = "a4",
+                     version = "*",
+                     remote_provider = "")  # BIOC provider
+  testthat::expect_equal(a$sameAs, "https://bioconductor.org/packages/release/bioc/html/a4.html")
+
+
+  testthat::expect_equal(a$provider$`@id`, "https://www.bioconductor.org")
+
+  a <- format_depend(package = "httr",
+                     version = "*",
+                     remote_provider = "") # CRAN provider
+  testthat::expect_equal(a$sameAs, "https://CRAN.R-project.org/package=httr")
+
+  a <- format_depend(package = "desc",
+                     version = "*",
+                     remote_provider = "r-lib/desc") # CRAN provider
+  testthat::expect_equal(a$sameAs, "https://github.com/r-lib/desc")
+
+  f <- system.file("examples/DESCRIPTION_with_remote", package = "codemetar")
+  descr <- codemeta_description(f)
+  testthat::expect_equal(descr$softwareRequirements[6, 1][[1]],
+                         "https://github.com/hadley/readr")
+
+  testthat::expect_equal(a$provider$`@id`, "https://cran.r-project.org")
+  a <- format_depend(package = "R",
+                     version = ">= 3.0.0",
+                     remote_provider = "")
+  testthat::expect_equal(a$name, "R")
 
 })
 
 testthat::test_that("Test the various cases for ids (NOT used currently)", {
 
-  a <- guess_dep_id(parse_depends("a4")[[1]])  # BIOC provider
-  a <- guess_dep_id(parse_depends("httr")[[1]]) # CRAN provider
-  a <- guess_dep_id(parse_depends("R")[[1]])
-  a <- guess_dep_id(parse_depends("not-a-package")[[1]])
+  # a <- guess_dep_id(parse_depends("a4")[[1]])  # BIOC provider
+  # a <- guess_dep_id(parse_depends("httr")[[1]]) # CRAN provider
+  # a <- guess_dep_id(parse_depends("R")[[1]])
+  # a <- guess_dep_id(parse_depends("not-a-package")[[1]])
 
 })